% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\docType{package}
\name{adexchangebuyer_googleAuthR}
\alias{adexchangebuyer_googleAuthR}
\alias{adexchangebuyer_googleAuthR-package}
\title{Ad Exchange Buyer API
Accesses your bidding-account information, submits creatives for validation, finds available direct deals, and retrieves performance reports.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:30:38
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangebuyerv13.auto/R/adexchangebuyer_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}
}

