% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{BillingInfo}
\alias{BillingInfo}
\title{BillingInfo Object}
\usage{
BillingInfo(accountId = NULL, accountName = NULL, billingId = NULL)
}
\arguments{
\item{accountId}{Account id}

\item{accountName}{Account name}

\item{billingId}{A list of adgroup IDs associated with this particular account}
}
\value{
BillingInfo object
}
\description{
BillingInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration data for an Ad Exchange billing info.
}

