% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{BillingInfoList}
\alias{BillingInfoList}
\title{BillingInfoList Object}
\usage{
BillingInfoList(items = NULL)
}
\arguments{
\item{items}{A list of billing info relevant for your account}
}
\value{
BillingInfoList object
}
\description{
BillingInfoList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A billing info feed lists Billing Info the Ad Exchange buyer account has access to. Each entry in the feed corresponds to a single billing info.
}

