% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Budget}
\alias{Budget}
\title{Budget Object}
\usage{
Budget(accountId = NULL, billingId = NULL, budgetAmount = NULL,
  currencyCode = NULL, id = NULL)
}
\arguments{
\item{accountId}{The id of the account}

\item{billingId}{The billing id to determine which adgroup to provide budget information for}

\item{budgetAmount}{The daily budget amount in unit amount of the account currency to apply for the billingId provided}

\item{currencyCode}{The currency code for the buyer}

\item{id}{The unique id that describes this item}
}
\value{
Budget object
}
\description{
Budget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration data for Ad Exchange RTB - Budget API.
}
\seealso{
Other Budget functions: \code{\link{budget.patch}},
  \code{\link{budget.update}}
}

