% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.corrections}
\alias{Creative.corrections}
\title{Creative.corrections Object}
\usage{
Creative.corrections()
}
\value{
Creative.corrections object
}
\description{
Creative.corrections Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Creative functions: \code{\link{Creative.disapprovalReasons}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}}, \code{\link{Creative}},
  \code{\link{creatives.insert}}
}

