% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.disapprovalReasons}
\alias{Creative.disapprovalReasons}
\title{Creative.disapprovalReasons Object}
\usage{
Creative.disapprovalReasons()
}
\value{
Creative.disapprovalReasons object
}
\description{
Creative.disapprovalReasons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}}, \code{\link{Creative}},
  \code{\link{creatives.insert}}
}

