% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.filteringReasons}
\alias{Creative.filteringReasons}
\title{Creative.filteringReasons Object}
\usage{
Creative.filteringReasons(Creative.filteringReasons.reasons = NULL,
  date = NULL, reasons = NULL)
}
\arguments{
\item{Creative.filteringReasons.reasons}{The \link{Creative.filteringReasons.reasons} object or list of objects}

\item{date}{The date in ISO 8601 format for the data}

\item{reasons}{The filtering reasons}
}
\value{
Creative.filteringReasons object
}
\description{
Creative.filteringReasons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The filtering reasons for the creative. Read-only. This field should not be set in requests.
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections}},
  \code{\link{Creative.disapprovalReasons}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd.logo}},
  \code{\link{Creative.nativeAd}}, \code{\link{Creative}},
  \code{\link{creatives.insert}}
}
