% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{Creative.nativeAd.logo}
\alias{Creative.nativeAd.logo}
\title{Creative.nativeAd.logo Object}
\usage{
Creative.nativeAd.logo(height = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{No description}

\item{url}{No description}

\item{width}{No description}
}
\value{
Creative.nativeAd.logo object
}
\description{
Creative.nativeAd.logo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A smaller image, for the advertiser logo.
}
\seealso{
Other Creative functions: \code{\link{Creative.corrections}},
  \code{\link{Creative.disapprovalReasons}},
  \code{\link{Creative.filteringReasons.reasons}},
  \code{\link{Creative.filteringReasons}},
  \code{\link{Creative.nativeAd.appIcon}},
  \code{\link{Creative.nativeAd.image}},
  \code{\link{Creative.nativeAd}}, \code{\link{Creative}},
  \code{\link{creatives.insert}}
}
