% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{CreativesList}
\alias{CreativesList}
\title{CreativesList Object}
\usage{
CreativesList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{A list of creatives}

\item{nextPageToken}{Continuation token used to page through creatives}
}
\value{
CreativesList object
}
\description{
CreativesList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The creatives feed lists the active creatives for the Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single creative.
}

