% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{DirectDealsList}
\alias{DirectDealsList}
\title{DirectDealsList Object}
\usage{
DirectDealsList(directDeals = NULL)
}
\arguments{
\item{directDeals}{A list of direct deals relevant for your account}
}
\value{
DirectDealsList object
}
\description{
DirectDealsList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A direct deals feed lists Direct Deals the Ad Exchange buyer account has access to. This includes direct deals set up for the buyer account as well as its merged stream seats.
}

