% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PerformanceReport}
\alias{PerformanceReport}
\title{PerformanceReport Object}
\usage{
PerformanceReport(bidRate = NULL, bidRequestRate = NULL,
  calloutStatusRate = NULL, cookieMatcherStatusRate = NULL,
  creativeStatusRate = NULL, filteredBidRate = NULL,
  hostedMatchStatusRate = NULL, inventoryMatchRate = NULL,
  latency50thPercentile = NULL, latency85thPercentile = NULL,
  latency95thPercentile = NULL, noQuotaInRegion = NULL, outOfQuota = NULL,
  pixelMatchRequests = NULL, pixelMatchResponses = NULL,
  quotaConfiguredLimit = NULL, quotaThrottledLimit = NULL, region = NULL,
  successfulRequestRate = NULL, timestamp = NULL,
  unsuccessfulRequestRate = NULL)
}
\arguments{
\item{bidRate}{The number of bid responses with an ad}

\item{bidRequestRate}{The number of bid requests sent to your bidder}

\item{calloutStatusRate}{Rate of various prefiltering statuses per match}

\item{cookieMatcherStatusRate}{Average QPS for cookie matcher operations}

\item{creativeStatusRate}{Rate of ads with a given status}

\item{filteredBidRate}{The number of bid responses that were filtered due to a policy violation or other errors}

\item{hostedMatchStatusRate}{Average QPS for hosted match operations}

\item{inventoryMatchRate}{The number of potential queries based on your pretargeting settings}

\item{latency50thPercentile}{The 50th percentile round trip latency(ms) as perceived from Google servers for the duration period covered by the report}

\item{latency85thPercentile}{The 85th percentile round trip latency(ms) as perceived from Google servers for the duration period covered by the report}

\item{latency95thPercentile}{The 95th percentile round trip latency(ms) as perceived from Google servers for the duration period covered by the report}

\item{noQuotaInRegion}{Rate of various quota account statuses per quota check}

\item{outOfQuota}{Rate of various quota account statuses per quota check}

\item{pixelMatchRequests}{Average QPS for pixel match requests from clients}

\item{pixelMatchResponses}{Average QPS for pixel match responses from clients}

\item{quotaConfiguredLimit}{The configured quota limits for this account}

\item{quotaThrottledLimit}{The throttled quota limits for this account}

\item{region}{The trading location of this data}

\item{successfulRequestRate}{The number of properly formed bid responses received by our servers within the deadline}

\item{timestamp}{The unix timestamp of the starting time of this performance data}

\item{unsuccessfulRequestRate}{The number of bid responses that were unsuccessful due to timeouts, incorrect formatting, etc}
}
\value{
PerformanceReport object
}
\description{
PerformanceReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration data for an Ad Exchange performance report list.
}
