% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PerformanceReportList}
\alias{PerformanceReportList}
\title{PerformanceReportList Object}
\usage{
PerformanceReportList(performanceReport = NULL)
}
\arguments{
\item{performanceReport}{A list of performance reports relevant for the account}
}
\value{
PerformanceReportList object
}
\description{
PerformanceReportList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The configuration data for an Ad Exchange performance report list.
}
