% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_functions.R
\name{pretargetingConfig.delete}
\alias{pretargetingConfig.delete}
\title{Deletes an existing pretargeting config.}
\usage{
pretargetingConfig.delete(accountId, configId)
}
\arguments{
\item{accountId}{The account id to delete the pretargeting config for}

\item{configId}{The specific id of the configuration to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest}{Google Documentation}
}

