% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PretargetingConfig.dimensions}
\alias{PretargetingConfig.dimensions}
\title{PretargetingConfig.dimensions Object}
\usage{
PretargetingConfig.dimensions()
}
\value{
PretargetingConfig.dimensions object
}
\description{
PretargetingConfig.dimensions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PretargetingConfig functions: \code{\link{PretargetingConfig.excludedPlacements}},
  \code{\link{PretargetingConfig.placements}},
  \code{\link{PretargetingConfig}},
  \code{\link{pretargetingConfig.insert}},
  \code{\link{pretargetingConfig.patch}},
  \code{\link{pretargetingConfig.update}}
}
