% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer_objects.R
\name{PretargetingConfigList}
\alias{PretargetingConfigList}
\title{PretargetingConfigList Object}
\usage{
PretargetingConfigList(items = NULL)
}
\arguments{
\item{items}{A list of pretargeting configs}
}
\value{
PretargetingConfigList object
}
\description{
PretargetingConfigList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

